flaskApp/requirements.txt
Flask==2.2.3
psycopg2==2.9.6
azure-storage-blob==12.10.0
python-dotenv==0.21.1